+++ NEWS.md
@@ -1,6 +1,7 @@
 # santoku (development version)
 
 * Add `brk_equally()` for symmetry with `chop_equally()`.
+* Experimental: `lbl_integer()` for integer data.
 * Minor tweaks to `chop_deciles()`.
 * Bugfix: `brk_left/right()` only affect specified breaks, not results of
   `extend`, when used as wrappers.